+++ src/actions/ShellAction/selectShell.ts
@@ -1,5 +1,4 @@
-import * as os from "os";
-import { OperatingSystemType } from "../../OperatingSystem";
+import { CURRENT_OS_TYPE } from "../../OperatingSystem";
 import { ShellType } from "./IShellActionConfig";
 import { IShellInvoker } from "./IShellInvoker";
 import { ShellInvokerPowerShell } from "./ShellInvokerPowerShell";
@@ -14,8 +13,7 @@ export function selectShell(shellType: ShellType | undefined): IShellInvoker {
             return new ShellInvokerSH();
 
         case undefined:
-            const osType: OperatingSystemType = os.type() as OperatingSystemType;
-            return (osType === "Windows_NT")      // tslint:disable-line:newline-before-return
+            return (CURRENT_OS_TYPE === "windows")
                 ? new ShellInvokerPowerShell()
                 : new ShellInvokerSH();