+++ test/LinkAction.spec.ts
@@ -8,6 +8,7 @@ import { ILinkCreationConfig } from "../src/actions/LinkAction/ILinkCreationConf
 import { ILinkCreator } from "../src/actions/LinkAction/ILinkCreator";
 import { LinkAction } from "../src/actions/LinkAction/LinkAction";
 import { ILogger } from "../src/Logger/ILogger";
+import { CURRENT_OS_TYPE } from "../src/OperatingSystem";
 
 describe("LinkAction", () => {
 
@@ -64,17 +65,19 @@ describe("LinkAction", () => {
 
     // tslint:disable-next-line:typedef
     const platformLinks = {
-      Linux: expectedLinksLinux,
-      Windows_NT: expectedLinksWindows,
-      Darwin: expectedLinksOsx,
+      linux: expectedLinksLinux,
+      windows: expectedLinksWindows,
+      osx: expectedLinksOsx,
     };
 
     /* tslint:disable:typedef */
     const loggerMock = TypeMoq.Mock.ofType<ILogger>();
     loggerMock.setup(m => m.logInfo(TypeMoq.It.is(s => s.startsWith("  [LINK]")))).verifiable(TypeMoq.Times.exactly(3));
 
+    const expectedOSLinks: Map<string, string[]> = platformLinks[CURRENT_OS_TYPE];
+
     const linkCreatorMock = TypeMoq.Mock.ofType<ILinkCreator>(undefined, TypeMoq.MockBehavior.Strict);
-    expectedLinksOsx.forEach((destinationPaths: string[], targetPath: string) => {
+    expectedOSLinks.forEach((destinationPaths: string[], targetPath: string) => {
       for (const destinationPath of destinationPaths) {
         linkCreatorMock.setup(
           async m => m.createLinkAsync(
@@ -98,11 +101,6 @@ describe("LinkAction", () => {
       }
     });
     /* tslint:enable:typedef */
-
-    // tslint:disable-next-line:no-require-imports typedef
-    const os = require("os");
-    // tslint:disable-next-line:no-unsafe-any
-    os.type = (): string => "Darwin";
 
     const linkAction: LinkAction = new LinkAction(
       expectedLinksGlobal,
@@ -136,9 +134,9 @@ describe("LinkAction", () => {
     const expectedLinks: Map<string, string[]> = new Map<string, string[]>();
     // tslint:disable-next-line:typedef
     const platformLinks = {
-      Darwin: new Map<string, string[]>(),
-      Windows_NT: new Map<string, string[]>(),
-      Linux: new Map<string, string[]>(),
+      osx: new Map<string, string[]>(),
+      windows: new Map<string, string[]>(),
+      linux: new Map<string, string[]>(),
     };
 
     // tslint:disable-next-line:no-require-imports typedef
@@ -258,9 +256,9 @@ describe("LinkAction", () => {
 
     // tslint:disable-next-line:typedef
     const platformLinks = {
-      Linux: expectedLinksLinux,
-      Windows_NT: expectedLinksWindows,
-      Darwin: expectedLinksOsx,
+      linux: expectedLinksLinux,
+      windows: expectedLinksWindows,
+      osx: expectedLinksOsx,
     };
 
     const expectedLinkAction: LinkAction = new LinkAction(expectedLinksGlobal, platformLinks, expectedLinkConfig);